goversion: Split Describe() off from main()-level concerns
